backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.54.0
requests==2.31.0
geopy==2.4.1
python-dotenv==1.0.0
bcrypt==4.1.2
pyjwt==2.8.0
